Fix compile warning
M  src/insexpand.c

+++ src/insexpand.c
@@ -5895,7 +5895,7 @@ find_common_prefix(size_t *prefix_len, int curbuf_only)
 
     vim_free(match_count);
 
-    if (len > ins_compl_leader_len())
+    if (len > (int)ins_compl_leader_len())
     {
 	*prefix_len = (size_t)len;
 	return first;
